Tidy up the motor control logic.
Using member variables for temporaries was very confusing. Replace them
with temporaries passed into the functions as required.

Replace drivePin and gndPin with a single Direction saying which way we
are driving; write_to_pins can choose which pins to drive based on that.
I am assuming that coasting is the right thing to do if 0 PWM is
requested.

Encoder.revolutions now returns a signed value so there is no need to
multiply by the direction we think we are moving.