Don't allow implicit numeric conversions.
The code was performing implicit numeric conversions (between float and
double, and float and int) all the time. Float to int conversions lose
precision; float to double promotions often mean the compiler is forced
to work in doubles, which are very slow on the itsybitsy.

Use a GCC #prgama to make all implicit conversions into errors, and tidy
up the code to conform to this. It is annoying that the Twist message
uses doubles, especially since it is actually converted to and from
single-precision before being sent over the serial line. But that can't
be fixed without patching rosserial_arduino.